+++ src/main/java/io/supertokens/storage/postgresql/queries/GeneralQueries.java
@@ -475,6 +475,9 @@ public class GeneralQueries {
 
     @TestOnly
     public static void deleteAllTables(Start start) throws SQLException, StorageQueryException {
+        if (!Start.isTesting) {
+            throw new UnsupportedOperationException();
+        }
         {
             String DROP_QUERY = "DROP INDEX IF EXISTS emailpassword_password_reset_token_expiry_index";
             update(start, DROP_QUERY, NO_OP_SETTER);
@@ -1069,6 +1072,51 @@ public class GeneralQueries {
         }
 
         return new HashMap<>();
+    }
+
+    @TestOnly
+    public static String[] getAllTablesInTheDatabase(Start start) throws StorageQueryException, SQLException {
+        if (!Start.isTesting) {
+            throw new UnsupportedOperationException();
+        }
+        String QUERY = "SELECT tablename, schemaname FROM pg_catalog.pg_tables "
+                + " WHERE schemaname = ?";
+
+        return execute(start, QUERY, pst -> {
+            pst.setString(1, Config.getConfig(start).getTableSchema());
+        }, result -> {
+            List<String> tableNames = new ArrayList<>();
+
+            while (result.next()) {
+                tableNames.add(result.getString("tablename"));
+            }
+            return tableNames.toArray(new String[0]);
+        });
+    }
+
+    @TestOnly
+    public static String[] getAllTablesInTheDatabaseThatHasDataForAppId(Start start, String appId)
+            throws StorageQueryException, SQLException {
+        if (!Start.isTesting) {
+            throw new UnsupportedOperationException();
+        }
+        String[] tableNames = getAllTablesInTheDatabase(start);
+
+        List<String> result = new ArrayList<>();
+        for (String tableName : tableNames) {
+            String QUERY = "SELECT 1 FROM " + Config.getConfig(start).getTableSchema() + "." + tableName + " WHERE app_id = ?";
+
+            boolean hasRows = execute(start, QUERY, pst -> {
+                pst.setString(1, appId);
+            }, res -> {
+                 return res.next();
+            });
+            if (hasRows) {
+                result.add(tableName);
+            }
+        }
+
+        return result.toArray(new String[0]);
     }
 
     private static class UserInfoPaginationResultHolder {
